Move nowcast abstractions into data pipeline
