+++ pipelines/data/nowcast.py
@@ -1,12 +1,12 @@
-"""
-Generic nowcast objects for EpiAutoGP.
-"""
+"""Generic nowcast data and source abstractions."""
+
+from __future__ import annotations
 
 import datetime as dt
 from dataclasses import dataclass, field
 from typing import Protocol
 
-from pipelines.epiautogp.forecast_spec import ForecastSpec
+from pipelines.data.forecast_spec import ForecastSpec
 
 
 @dataclass(frozen=True)
