+++ pipelines/tests/epiautogp/test_prep_epiautogp_data.py
@@ -15,12 +15,12 @@ import polars as pl
 import pytest
 
 from pipelines.data.hubverse_nowcast import HubverseNowcast
+from pipelines.data.nowcast import NowcastData
 from pipelines.epiautogp.epiautogp_forecast_utils import (
     ForecastPipelineContext,
     ForecastSpec,
     ModelPaths,
 )
-from pipelines.epiautogp.nowcast import NowcastData
 from pipelines.epiautogp.prep_epiautogp_data import (
     _apply_date_exclusions,
     convert_to_epiautogp_json,
